add image to index
